changed model to support full publisher.  Added new logging

+++ functions/streamtoidv/clean.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+if [ -z "$CLEAN" ]; then
+    echo "No clean flag set exiting."
+    exit 0
+fi
 
 ### Cleanup virtualenv post-deploy
 find . -d -not -name ".." -not -name "." -not -name "*.json" -not -name "clean.sh" -not -name "main.py" -not -name "function.*" -not -name "utils.py" -not -name "requirements.txt" | xargs rm -fr &> /dev/null
